Divide Sku class
