+++ sdk/resourcemanager/Azure.ResourceManager.Core/src/Resources/Sku.cs
@@ -2,16 +2,13 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Text.Json;
-using Azure.Core;
-using Azure.ResourceManager.Resources.Models;
 
 namespace Azure.ResourceManager.Core
 {
     /// <summary>
-    /// A class representing SKU for resource
+    /// A class representing SKU for resource.
     /// </summary>
-    public sealed class Sku : IEquatable<Sku>, IComparable<Sku>
+    public sealed partial class Sku : IEquatable<Sku>, IComparable<Sku>
     {
         /// <summary>
         /// Initializes a new instance of the <see cref="Sku"/> class.
@@ -107,101 +104,6 @@ namespace Azure.ResourceManager.Core
                 string.Equals(Size, other.Size, StringComparison.InvariantCultureIgnoreCase) &&
                 string.Equals(Tier, other.Tier, StringComparison.InvariantCultureIgnoreCase) &&
                 long.Equals(Capacity, other.Capacity);
-        }
-
-        /// <summary>
-        /// Serialize the input Sku object.
-        /// </summary>
-        /// <param name="writer"> Input Json writer. </param>
-        /// <param name="value"> Input Sku object. </param>
-        internal static void Serialize(Utf8JsonWriter writer, Sku value)
-        {
-            if (writer is null)
-            {
-                throw new ArgumentNullException(nameof(writer));
-            }
-
-            writer.WriteStartObject();
-            if (Optional.IsDefined(value.Name))
-            {
-                writer.WritePropertyName("name");
-                writer.WriteStringValue(value.Name);
-            }
-            if (Optional.IsDefined(value.Tier))
-            {
-                writer.WritePropertyName("tier");
-                writer.WriteStringValue(value.Tier);
-            }
-            if (Optional.IsDefined(value.Size))
-            {
-                writer.WritePropertyName("size");
-                writer.WriteStringValue(value.Size);
-            }
-            if (Optional.IsDefined(value.Family))
-            {
-                writer.WritePropertyName("family");
-                writer.WriteStringValue(value.Family);
-            }
-            if (Optional.IsDefined(value.Capacity))
-            {
-                writer.WritePropertyName("capacity");
-                writer.WriteNumberValue(value.Capacity.Value);
-            }
-            writer.WriteEndObject();
-        }
-
-        /// <summary>
-        /// Deserialize the input Json object.
-        /// </summary>
-        /// <param name="element"> The Json object need to be deserialized. </param>
-        internal static Sku Deserialize(JsonElement element)
-        {
-            Optional<string> name = default;
-            Optional<string> tier = default;
-            Optional<string> size = default;
-            Optional<string> family = default;
-            Optional<string> model = default;
-            Optional<int> capacity = default;
-            foreach (var property in element.EnumerateObject())
-            {
-                if (property.NameEquals("name"))
-                {
-                    name = property.Value.GetString();
-                    continue;
-                }
-                if (property.NameEquals("tier"))
-                {
-                    tier = property.Value.GetString();
-                    continue;
-                }
-                if (property.NameEquals("size"))
-                {
-                    size = property.Value.GetString();
-                    continue;
-                }
-                if (property.NameEquals("family"))
-                {
-                    family = property.Value.GetString();
-                    continue;
-                }
-                if (property.NameEquals("model"))
-                {
-                    model = property.Value.GetString();
-                    continue;
-                }
-                if (property.NameEquals("capacity"))
-                {
-                    if (property.Value.ValueKind == JsonValueKind.Null)
-                    {
-                        property.ThrowNonNullablePropertyIsNull();
-                        continue;
-                    }
-                    capacity = property.Value.GetInt32();
-                    continue;
-                }
-            }
-
-            return new Sku(name.Value, tier.Value, family.Value, size.Value, Optional.ToNullable(capacity));
         }
     }
 }
